refactor: useUserViewModelをModelに統合して削除
